🏓🌑 Checkpoint ./komutlar/kayıtsayı-sıfırla.js:26490528/328 ./komutlar/isimler.js:26490528/22 ./komutlar/kayıtsayı.js:26490528/6

+++ komutlar/kayıtsayı-sıfırla.js
@@ -11,10 +11,11 @@ if (message.author.id !== ayarlar.sahip & message.author.id !== ayarlar.baran & 
 return message.channel.send(new MessageEmbed().setDescription(`${emoji} ${message.author}, Bu işlemi sadece yetkililer yapabilir`).setAuthor(message.member.displayName, message.author.avatarURL({ dynamic: true })).setColor('0x800d0d').setTimestamp()).then(x => x.delete({timeout: 5000}))
 
 const member = message.guild.member(message.mentions.members.first() || message.guild.members.cache.get(args[0]));
-let üye = message.mentions.users.first() || message.author
+let üye = message.mentions.users.first()
+if(!üye) return message.channel.send(new MessageEmbed().setDescription(`${emoji} ${message.author}, Bir kullanıcı etiketlemelisin.`).setAuthor(message.member.displayName, message.author.avatarURL({ dynamic: true })).setColor('0x800d0d').setTimestamp()).then(x => x.delete({timeout: 5000}));
 
 db.delete(`erkek.sayı_${üye.id}`)
-db.delete(`toplam.sayı_${message.author.id}`)  
+db.delete(`toplam.sayı_${üye.id}`)  
 db.delete(`kadın.sayı_${üye.id}`) 
 message.react(basari)
